Rename config_macros and configMacros to configmacros
